api/requirements.txt
flask[async]
flask-cors
supabase
python-dotenv
pydantic
